feat(artBox): add logic to ajouterPost handler
Adds the logic in charge to verify that the form is correctly filed
and if so adds the new `oeuvre` in db and redirects to the home page.

If the form isn't correctly filed the ajouter html page is display
with an error message with each field that didn't pass the validation.